29.12.2020

+++ database/migrations/2020_02_05_071222_create_categories_table.php
@@ -24,7 +24,7 @@ class CreateCategoriesTable extends Migration
             $table->text('description')->nullable();
             $table->text('body')->nullable();
             $table->string('img')->nullable()->default(config('admin.imgProductDefault'));
-            $table->integer('popular')->unsigned()->default('0');
+            $table->bigInteger('popular')->unsigned()->default('0');
             $table->softDeletes();
             $table->timestamps();
         });
